requirements.txt
# Core ML
lightgbm>=4.0
xgboost>=2.0
catboost>=1.2
scikit-learn>=1.3
torch>=2.0

# Data & Features
pandas>=2.0
numpy>=1.24
pandas-ta>=0.3

# Validation & Financial ML
mlfinlab>=2.0

# Backtesting & Analysis
vectorbt>=0.26
quantstats>=0.0.62

# Regime Detection
hmmlearn>=0.3

# Optimization
optuna>=3.0

# MT5 (Windows only)
MetaTrader5>=5.0

# Utilities
joblib>=1.3
schedule>=1.2
python-dotenv>=1.0
PyYAML>=6.0
